Added GetNifDefNames()
Instead of looping through the child defs of the input element, as you might expect, this function instead loops through the child elements. This works because all elements exist by default, and cannot be deleted.

Not all elements are enabled, depending on the version of the Nif file. This is why the function loops through the child elements, instead of the child defs. Determining whether or not a def is enabled in the file sometimes requires you to pass the actual element itself (instead of just, say, a parent element). Resolving unions has the same problem. Instead of trying to rewrite parts of xEdit to "fix" this, it's much simpler to just loop through the child elements.

The exception to all of this is arrays, since they can have an arbitrary amount of elements which can be deleted and created. Therefore, if the input element is an array, the function returns the name of the first child def (arrays always have exactly 1 child def). No enabled check is done when the input is an array - if the array element itself is enabled, so are the children.