Removed filePath parameter from CreateNif()
Makes the function much more convenient to call in what I expect to be almost the only use case. There isn't a good to immediately save the empty nif to disk. You'll have to save it again, anyway, when you're done editing it.

Also added the missing nif version test.